move device to own file model
